refactor(ui): split built-in theme definitions
