+++ src/ui/themes/ember.ts
@@ -1,0 +1,53 @@
+import { withLazySyntaxStyle } from "./syntax";
+import type { AppTheme } from "./types";
+
+/** Warm dark theme with ember-like reds and oranges. */
+export const EMBER_THEME: AppTheme = withLazySyntaxStyle(
+  {
+    id: "ember",
+    label: "Ember",
+    appearance: "dark",
+    background: "#140b08",
+    panel: "#22120d",
+    panelAlt: "#2c1710",
+    border: "#643627",
+    accent: "#ffb07a",
+    accentMuted: "#5d3428",
+    text: "#fff0e6",
+    muted: "#c7a18d",
+    addedBg: "#183424",
+    removedBg: "#4a1f1f",
+    contextBg: "#24140e",
+    addedContentBg: "#21432c",
+    removedContentBg: "#5a2727",
+    contextContentBg: "#2b1711",
+    addedSignColor: "#83d99d",
+    removedSignColor: "#ff9d8f",
+    lineNumberBg: "#1c100c",
+    lineNumberFg: "#9a735f",
+    selectedHunk: "#8a4d3a",
+    badgeAdded: "#83d99d",
+    badgeRemoved: "#ff9d8f",
+    badgeNeutral: "#f1be9d",
+    fileNew: "#83d99d",
+    fileDeleted: "#ff9d8f",
+    fileRenamed: "#ffd08f",
+    fileModified: "#d8b4fe",
+    fileUntracked: "#ffb07a",
+    noteBorder: "#e1a3ff",
+    noteBackground: "#311d36",
+    noteTitleBackground: "#452650",
+    noteTitleText: "#fff0ff",
+  },
+  {
+    default: "#fff0e6",
+    keyword: "#ffb47f",
+    string: "#ffd3a8",
+    comment: "#a17d69",
+    number: "#ffd08f",
+    function: "#ffd9b3",
+    property: "#ffc89f",
+    type: "#f7c5b0",
+    punctuation: "#a17d69",
+  },
+);
