+++ src/ui/themes/graphite.ts
@@ -1,0 +1,53 @@
+import { withLazySyntaxStyle } from "./syntax";
+import type { AppTheme } from "./types";
+
+/** Default dark theme with a neutral graphite palette. */
+export const GRAPHITE_THEME: AppTheme = withLazySyntaxStyle(
+  {
+    id: "graphite",
+    label: "Graphite",
+    appearance: "dark",
+    background: "#111315",
+    panel: "#171a1d",
+    panelAlt: "#1d2126",
+    border: "#343c45",
+    accent: "#d5e0ea",
+    accentMuted: "#414a54",
+    text: "#f2f4f6",
+    muted: "#9aa4af",
+    addedBg: "#1f3025",
+    removedBg: "#372526",
+    contextBg: "#181c20",
+    addedContentBg: "#24362a",
+    removedContentBg: "#432b2d",
+    contextContentBg: "#1e2328",
+    addedSignColor: "#88d39b",
+    removedSignColor: "#f0a0a0",
+    lineNumberBg: "#14181b",
+    lineNumberFg: "#798592",
+    selectedHunk: "#4f5d6b",
+    badgeAdded: "#88d39b",
+    badgeRemoved: "#f0a0a0",
+    badgeNeutral: "#a9b4bf",
+    fileNew: "#88d39b",
+    fileDeleted: "#f0a0a0",
+    fileRenamed: "#e6cf98",
+    fileModified: "#c49bff",
+    fileUntracked: "#7fd1ff",
+    noteBorder: "#c6a0ff",
+    noteBackground: "#241c31",
+    noteTitleBackground: "#322446",
+    noteTitleText: "#f5edff",
+  },
+  {
+    default: "#f2f4f6",
+    keyword: "#c4d0da",
+    string: "#d8c6ef",
+    comment: "#7f8b97",
+    number: "#e6cf98",
+    function: "#dfe6ed",
+    property: "#bac8d4",
+    type: "#d3d9e2",
+    punctuation: "#7f8b97",
+  },
+);
