+++ src/ui/themes/midnight.ts
@@ -1,0 +1,53 @@
+import { withLazySyntaxStyle } from "./syntax";
+import type { AppTheme } from "./types";
+
+/** Cool dark theme optimized for blue-toned terminals. */
+export const MIDNIGHT_THEME: AppTheme = withLazySyntaxStyle(
+  {
+    id: "midnight",
+    label: "Midnight",
+    appearance: "dark",
+    background: "#08111f",
+    panel: "#0e1b2e",
+    panelAlt: "#13243a",
+    border: "#284264",
+    accent: "#7fd1ff",
+    accentMuted: "#355578",
+    text: "#eef4ff",
+    muted: "#8da5c7",
+    addedBg: "#153526",
+    removedBg: "#47262a",
+    contextBg: "#0f1b2d",
+    addedContentBg: "#102a1f",
+    removedContentBg: "#371b1e",
+    contextContentBg: "#132238",
+    addedSignColor: "#69d69a",
+    removedSignColor: "#ff8e8e",
+    lineNumberBg: "#0b1627",
+    lineNumberFg: "#56739a",
+    selectedHunk: "#2a6a8a",
+    badgeAdded: "#5ad188",
+    badgeRemoved: "#ff8b8b",
+    badgeNeutral: "#89a5d3",
+    fileNew: "#5ad188",
+    fileDeleted: "#ff8b8b",
+    fileRenamed: "#ffd883",
+    fileModified: "#b794f6",
+    fileUntracked: "#7fd1ff",
+    noteBorder: "#c49bff",
+    noteBackground: "#211a36",
+    noteTitleBackground: "#30234f",
+    noteTitleText: "#f5eeff",
+  },
+  {
+    default: "#e8f1ff",
+    keyword: "#8ed4ff",
+    string: "#c7b4ff",
+    comment: "#6e85a7",
+    number: "#ffd883",
+    function: "#b6c9ff",
+    property: "#a8d6ff",
+    type: "#a4b7ff",
+    punctuation: "#6e85a7",
+  },
+);
