+++ src/ui/themes/paper.ts
@@ -1,0 +1,53 @@
+import { withLazySyntaxStyle } from "./syntax";
+import type { AppTheme } from "./types";
+
+/** Warm light theme with paper-inspired neutrals. */
+export const PAPER_THEME: AppTheme = withLazySyntaxStyle(
+  {
+    id: "paper",
+    label: "Paper",
+    appearance: "light",
+    background: "#f4efe6",
+    panel: "#fffaf3",
+    panelAlt: "#f8f1e7",
+    border: "#d8c8b3",
+    accent: "#77593a",
+    accentMuted: "#d7ccbe",
+    text: "#2f2417",
+    muted: "#786753",
+    addedBg: "#dff0e1",
+    removedBg: "#f6ddde",
+    contextBg: "#faf6ee",
+    addedContentBg: "#eaf8ec",
+    removedContentBg: "#fbebeb",
+    contextContentBg: "#fffaf3",
+    addedSignColor: "#3f8d58",
+    removedSignColor: "#b4545b",
+    lineNumberBg: "#f2e9dc",
+    lineNumberFg: "#9b8367",
+    selectedHunk: "#d2c0a5",
+    badgeAdded: "#3f8d58",
+    badgeRemoved: "#b4545b",
+    badgeNeutral: "#8e7355",
+    fileNew: "#3f8d58",
+    fileDeleted: "#b4545b",
+    fileRenamed: "#9f6c1f",
+    fileModified: "#7d5bc4",
+    fileUntracked: "#4a6890",
+    noteBorder: "#7d5bc4",
+    noteBackground: "#efe6ff",
+    noteTitleBackground: "#e3d7ff",
+    noteTitleText: "#462b74",
+  },
+  {
+    default: "#2f2417",
+    keyword: "#7b5a35",
+    string: "#4a6890",
+    comment: "#8f7a65",
+    number: "#9f6c1f",
+    function: "#5a4a8e",
+    property: "#356b7f",
+    type: "#5f5f9a",
+    punctuation: "#8f7a65",
+  },
+);
